requirements.txt
# Core dependencies
numpy==1.26.4
opencv-python==4.11.0.86
mediapipe==0.10.21
torch==2.3.0
torchvision==0.18.0

# Data processing and analysis
pandas==2.2.0
scikit-learn==1.4.0
matplotlib==3.10.0
seaborn==0.13.0
plotly==5.17.0

# Configuration and utilities
pyyaml==6.0.1
hydra-core==1.3.2
omegaconf==2.3.0
tqdm==4.66.1
click==8.1.7
protobuf>=4.25.3,<5.0.0

# Logging and monitoring
wandb==0.16.1
tensorboard==2.16.2
loguru==0.7.2

# Testing and development
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Jupyter and notebooks
jupyter==1.0.0
ipykernel==6.26.0
ipywidgets==8.1.1

# Image processing and augmentation
albumentations==1.3.1
Pillow==10.1.0

# Performance optimization
numba==0.58.1
psutil==5.9.6

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
